requirements.txt
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
Django==5.0.2
google-auth==2.36.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
numpy==1.24.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
requests==2.32.3
rsa==4.9
sniffio==1.3.1
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.8.2
google-generativeai==0.3.2
python-dotenv==1.0.1
pinecone-client==3.0.2
sentence-transformers==2.5.1
torch==2.2.1
